Changes in byte string operations.
String.bytes property is changed to String.toBytes() method.
The method serializes an Unicode string to a byte string.
It returns null if a character larger than 255 is encountered
in the Unicode string.
String.fromBytes() method converts a byte string to an Unicode string.
String.utf8 property is converted to String.fromUTF8() method.
The method converts an UTF-8 encoded byte string to an Unicode string.
String.toUTF8() serializes Unicode string to an UTF-8 encoded byte string.