Fixed passing uninitialized value to printf().
Previously, njs_ext_console_log() passed an uninitialized value to
printf() function in the case of an error. The precison argument was set
to 0 in such a case, so well-behaved implementation was able to handle
it properly.

Setting this value to NULL, just to make sure and to make Coverity Scan
happy.