Making build and test targets separate.
Previously, if libedit was unavailable the default action was to build
libnjs and run tests. However, if it was available the default action
was just building binaries.

This patch makes the two high-level targets which are intended to be
invoked externally.  The default one for building all available binaries
and the test target for running all available tests.