Fixed index generation for global objects.
In c75a8fc6d534 "GLOBAL GET" instruction was introduced to handle unresolved
references.  The issue was that the "GLOBAL GET" instruction erroneously
used the assignment variable index as a destination index.
The result was that a variable was assigned the retval of a "GLOBAL GET"
instruction.

The fix is to use a separate temporary index for "GLOBAL GET".

This closes #289 issue on GitHub.