Modules: renaming vars to rawVariables to better reflect purpose.
In 434f20c29f4c, obj.vars was introduced.  obj.vars is almost identical
to obj.variables except a value of the Buffer type is returned.

Since most nginx variables are valid strings, it is preferable to
leave both variants.

To avoid confusion rawVariables name is used for Buffer variables.