Fixed template literal from producing byte-strings.
Previously, as a side effect of creating a key for the values hash a
byte-string was created.  This byte-string was reused internally and
might appear in template literal.  As a result a byte-string was
produced as a value for a template literal.  Byte-strings are obsolete
and are scheduled for removal because they can cause issues with
internal routines not prepared for them.