Change: native methods are provided with retval argument.
Previously, native methods were expected to return
their retval using vm->retval.  This caused problem in the part
(1aa137411b09, 293fe42c5e1c) because vm->retval can be overwritten
unexpectedly as a side-effect of operations like ToString(), ToNumber().

The fix is to never used a global retval. Instead methods
are provided with a retval argument to store their retval value.

As a part of the change, retval and exception values are split.
The normal value is returned in the retval argument.
The exception value is thrown by njs_vm_throw() or njs_vm_error().
The exception value can be acquired using njs_vm_exception_get().