feat(js_set): add nocache flag for uncacheable variables
This commit adds support for an additional `nocache` flag
in `js_set` directives. If set, the resulting nginx variable
will have no_cacheable set to 1. This enables us to dynamically
recompute a variable if the context changed (for example, in
case of an internal redirection).

In case of multiple calls in a location, users should cache the
result in a rewrite variable: `set $cached_variable $js_variable;`