auto/clang: Added a NJS_NONSTRING macro.
This is a wrapper around __attribute__((nonstring)). Traditionally this
was used to mark char array variables that intentionally lacked a
terminating NUL byte, this would then cause warnings to either be
quelled or emitted for various memory/string functions.

GCC 15 introduced a new warning, Wunterminated-string-initialization,
which will always warn on things like

  static const char hex[16] = "0123456789ABCDEF";

However this is very much intentionally not NUL terminated.

The above attribute also quells this new warning (which is also enabled
by -Wextra).

So the above example would become

  static const char hex[16] NJS_NONSTRING = "0123456789ABCDEF";

Link: <https://gcc.gnu.org/git/?p=gcc.git;a=commit;h=44c9403ed1833ae71a59e84f9e37af3182be0df5>
Link: <https://gcc.gnu.org/git/gitweb.cgi?p=gcc.git;h=622968990beee7499e951590258363545b4a3b57>
Link: <https://gcc.gnu.org/bugzilla/show_bug.cgi?id=117178#c21>
Signed-off-by: Andrew Clayton <a.clayton@nginx.com>